pyproject.toml
[project]
name = "palantirlite-demo-freelancer-finance"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "openpyxl>=3.1.5",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.2",
]
